requirements.txt
# --- Backend framework ---
Flask==2.2.5
flask-cors==3.0.10

# --- Database ---
pymongo>=4.5,<5
dnspython==2.4.2        # Required for MongoDB Atlas (SRV URI parsing)

# --- Environment variables ---
python-dotenv==1.0.0

# --- Data validation ---
pydantic==1.10.11

# --- AI & ML libraries ---
scikit-learn==1.2.2
numpy==1.24.2
pandas==1.5.3